Document behavioural environment variables
These environment variables controll the behavioural settings of the
server. E.g. if it has an API and if SSL shall be enabled for it. Not
included are settings that control the modalities of the service, e.g.
the amount of days after which the trash shall be emptied.
Those will be added in a second step.

This variables list was created by going through the configuration
files of the services.