Always write config files (add USE_OWN_SETTINGS)
Write the configuration files on every startup of the container as users
might have changed values of some or all env vars.
To still allow users to mount their own configuration files, we added
the USE_OWN_SETTINGS env var, that disables the modification of any
configuration file if set to true.