Allow configuration of ZoneMTA future date feature
ZoneMTA disables the future date feature by default. The future date
feature allows a message with a future date in the 'Date' header, to be
enqued for later sending. With this commit we enable that feature by
default and introduce a new environment variable that allows disabling
the feature again.